% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_cross_validation.R
\name{par_cross_validation}
\alias{par_cross_validation}
\title{Parallel Cross-validation for prophet}
\usage{
par_cross_validation(
  model,
  horizon,
  units,
  period = NULL,
  initial = NULL,
  cutoffs = NULL,
  n.jobs = -1L
)
}
\arguments{
\item{model}{Fitted Prophet model.}

\item{horizon}{Integer size of the horizon.}

\item{units}{String unit of the horizon, e.g., "days", "secs".}

\item{period}{Integer amount of time between cutoff dates. Same units as
horizon. If not provided, 0.5 * horizon is used.}

\item{initial}{Integer size of the first training period. If not provided,
3 * horizon is used. Same units as horizon.}

\item{cutoffs}{Vector of cutoff dates to be used during
cross-validation. If not provided works beginning from (end - horizon),
works backwards making cutoffs with a spacing of period until initial is
reached.}

\item{n.jobs}{\strong{(New)} Number of cores to run in parallel, if \code{-1} all available
cores are used (default is \code{-1})}
}
\value{
A dataframe with the forecast, actual value, and cutoff date.
}
\description{
Parallelized version of \link[prophet]{cross_validation} from prophet forecasting tool
}
\examples{
\dontrun{
  df <- data.frame(
    ds = seq.Date(as.Date('2022-01-01'), as.Date('2022-12-31'), 'day'),
    y = pi*cos(1:365)+rnorm(365))
  m <- prophet(df)
  cv <- par_cross_validation(
    model = m,
    horizon = 30,
    units = 'days',
    n.jobs = 2
  )
  print(cv)
}
}
